+++ Source/Identity.Fetcher/Identity.Domain/Clustering/Document.cs
@@ -10,6 +10,6 @@ namespace Identity.Domain.Clustering
         public string Title { get; set; }
         public string Description { get; set; }
         public DateTimeOffset Added { get; set; }
-        public int[] WordVector { get; set; }
+        public double[] WordVector { get; set; }
     }
 }
